+++ src/main/java/blackjack/domain/game/BlackjackGameResult.java
@@ -1,6 +1,7 @@
 package blackjack.domain.game;
 
 import blackjack.domain.player.Player;
+import blackjack.domain.player.Result;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
