+++ src/main/java/blackjack/domain/player/Players.java
@@ -1,7 +1,6 @@
 package blackjack.domain.player;
 
 import blackjack.domain.card.Deck;
-import blackjack.domain.game.Result;
 
 import java.util.*;
 import java.util.function.Function;
