+++ src/main/java/blackjack/domain/player/Result.java
@@ -1,4 +1,4 @@
-package blackjack.domain.game;
+package blackjack.domain.player;
 
 public enum Result {
     WIN,
